requirements.txt
labelbox
requests
python-dotenv
Pillow
pyyaml